requirements.txt
django
python-dotenv
psycopg2
djangorestframework
django-cors-headers
django-redis
selenium
numpy
beautifulsoup4
lxml
PyQt5